# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JobTrack API",
    "version": "1.0.0",
    "description": "JobTrack backend API documentation"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/events/{id}": {
      "patch": {
        "summary": "Update event",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete event",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EventId"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jobtrack_auth"
      }
    },
    "parameters": {
...
      "EventId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "schemas": {
...
      "EventType": {
        "type": "string",
        "enum": [
          "deadline",
          "result",
          "interview"
        ]
      },
...
      "Event": {
        "type": "object",
        "required": [
          "id",
          "application_id",
          "event_type",
          "scheduled_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "application_id": {
            "type": "string",
            "format": "uuid"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "interview_round": {
            "type": "integer",
            "nullable": true
          },
          "notified_d3": {
            "type": "boolean",
            "nullable": true
          },
          "notified_d1": {
            "type": "boolean",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "UpdateEventRequest": {
        "type": "object",
        "properties": {
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "interview_round": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    }
  }
}